requirements.txt
streamlit
google-cloud-firestore
numpy 
pandas
